✨ feat(extensions): convert whimsical to directory layout with tests
