+++ extensions/whimsical/helpers.ts
@@ -1,9 +1,4 @@
-/**
- * Original source: https://github.com/mitsuhiko/agent-stuff/blob/main/extensions/whimsical.ts
- */
-import type { ExtensionAPI } from "@mariozechner/pi-coding-agent";
-
-const messages = [
+export const messages = [
   // Short
   "Schlepping...",
   "Combobulating...",
@@ -462,16 +457,6 @@ const messages = [
   "Cherry-picking the commits...",
 ];
 
-function pickRandom(): string {
+export function pickRandom(): string {
   return messages[Math.floor(Math.random() * messages.length)];
-}
-
-export default function (pi: ExtensionAPI) {
-  pi.on("turn_start", async (_event, ctx) => {
-    ctx.ui.setWorkingMessage(pickRandom());
-  });
-
-  pi.on("turn_end", async (_event, ctx) => {
-    ctx.ui.setWorkingMessage(); // Reset for next time
-  });
 }
